Fix `PlayerDisconnect` hook being called with bogus player (#4501)
* Fix `PlayerDisconnect` hook being called with bogus player

One of the paths this hook is invoked from takes the following path:
- `Netplay.UpdateConnectedClients`
- `NetMessage.SyncDisconnectedPlayer`
- `NetMessage.SyncOnePlayer`
- `Player.Hooks.PlayerDisconnect`

During this path, just before the call to
`NetMessage.SyncDisconnectedPlayer`, the player's `RemoteClient` is
`Reset`. This causes the entire player object to be re-instantiated, and
the hook to then be invoked with this new (bogus) player.

To resolve this, we re-order the the call to `RemoteClient.Reset`
_after_ the `NetMessage.SyncDisconnectedPlayer` call, so the player
received hasn't yet been destroyed. We do need to set the
`RemoteClient.State` to `0` in lieu of this, to ensure that this player
is properly treated as disconnected.

* Clarify ModPlayer.PlayerDisconnect docs

---------

Co-authored-by: JavidPack <javidpack@gmail.com>